refactor: TransactionsView into smaller components
